outsourced date booking calculations to custom hook
