+++ src/components/form/DateInput.tsx
@@ -4,10 +4,7 @@ import { motion } from 'framer-motion';
 import 'react-datepicker/dist/react-datepicker.css';
 import css from './Input.module.css';
 import useErrorAnimation from '@/hooks/useErrorAnimation';
-import { useParams } from 'react-router-dom';
-import { useSelector } from 'react-redux';
-import { RootState } from '@/store/types';
-import Booking from '@/models/Booking';
+import useBookedDates from '@/hooks/useBookedDates';
 
 interface Errors {
   [key: string]: string;
@@ -19,39 +16,12 @@ interface DateProps {
   onUpdate: (id: string, value: string) => void;
 }
 
-function dateRange(from: string, till: string) {
-  const dates = [];
-  const startDate = new Date(from);
-  const endDate = new Date(till);
-
-  while (startDate <= endDate) {
-    dates.push(new Date(startDate));
-    startDate.setDate(startDate.getDate() + 1);
-  }
-
-  return dates;
-}
-
 const DateInput: React.FC<DateProps> = ({ id, errors, onUpdate }) => {
   const { value, updateFormData, x, delay, backgroundColor } = useErrorAnimation(id, errors, onUpdate);
-  const { shipId } = useParams();
-  console.log(shipId)
-
-  const users = useSelector((state: RootState) => state.users.users);
-
-  const bookedDates: Date[] = [];
-  users.forEach((user) => {
-    user.bookings.forEach((booking: Booking) => {
-      if (booking.id === shipId) {
-        const dates = dateRange(booking.from, booking.till);
-        bookedDates.push(...dates);
-      }
-    });
-  });
+  const bookedDates = useBookedDates();
 
   const today = new Date();
-  const maxDate = new Date(today);
-  maxDate.setDate(maxDate.getDate() + 30);
+  const maxDate = new Date(today.setDate(today.getDate() + 30));
 
   function changeHandler(date: Date) {
     updateFormData(date.toISOString());
